+++ meteor/server/migration/1_42_0.ts
@@ -2,9 +2,9 @@ import { addMigrationSteps } from './databaseMigration'
 import { StudioRouteSet, StudioRouteType } from '@sofie-automation/corelib/dist/dataModel/Studio'
 import { Studios } from '../collections'
 
-export const addSteps = addMigrationSteps('1.42.0', [
-	// Add some migrations!
+// Release 42
 
+export const addSteps = addMigrationSteps('1.42.0', [
 	{
 		id: 'Add new routeType property to routeSets where missing',
 		canBeRunAutomatically: true,
