+++ meteor/server/migration/1_46_0.ts
@@ -2,5 +2,5 @@ import { addMigrationSteps } from './databaseMigration'
 
 // Release 46
 export const addSteps = addMigrationSteps('1.46.0', [
-	// Add some migrations!
+	// >> No migration steps in this release <<
 ])
