+++ meteor/server/migration/migrations.ts
@@ -35,6 +35,9 @@ addSteps1_49_0()
 import { addSteps as addSteps1_50_0 } from './1_50_0'
 addSteps1_50_0()
 
+import { addSteps as addSteps1_51_0 } from './1_51_0'
+addSteps1_51_0()
+
 // Migrations for the in-development release:
 import { addSteps as addStepsX_X_X } from './X_X_X'
 addStepsX_X_X()
